Generalise run_test_obsforge_global_atmos_bufr_dump.sh and update SPOC (#170)
`test/atmos/run_test_obsforge_global_atmos_bufr_dump.sh` works well in
`ctests` (which is what it is for!) but can be more useful if expanded
to allow any date and also to relax the requirement that the test data
and code is in the same directory tree.

This PR does three things:

1) Allows an optional argument to specify the date (rather than picking
the most recent)
2) Changes the path used for `$topdir` calculation
3) Brings in the most recent version of
[spoc](https://github.com/NOAA-EMC/spoc/tree/17701f20531a6b8987ed35f2ba1bb7c749c63a3d)
that includes the new naming convention.

The relevent ctest still passes.

The script can be run standalone using the form:

```
./run_test_obsforge_global_atmos_bufr_dump.sh /scratch3/NCEPDEV/da/Andrew.Collard/ICSDIR /scratch3/NCEPDEV/da/Andrew.Collard/git/obsForge/build 2024022400
```